+++ examples/06-dev/demo/project/src/test/index/index.js
@@ -23,3 +23,9 @@ console.log('test/index');
 // test ignoreNodeModules
 const Swiper = require('swiper');
 require('bootstrap');
+
+const $ = require('jquery');
+
+$.get('/src/test/index/data/index', {}, res => {
+    console.log("get['/src/test/index/data/index']", res);
+});